Null check for null pointer exception in batch service

+++ src/main/java/com/revature/assignforce/service/BatchServiceImpl.java
@@ -49,6 +49,9 @@ public class BatchServiceImpl implements BatchService {
 	@Override
 	public void delete(int id) {
 		Batch batch = batchRepository.findById(id).get();
+		if ( batch == null ) {
+			return;
+		}
 		batch.setSkills(new HashSet<SkillIdHolder>());
 		batchRepository.save(batch);
 		batchRepository.deleteById(id);
